feat: fix TooltipTrigger styles (#692), :recycle: :lipstick:
