+++ src/pages/Options/Options.css
@@ -23,7 +23,7 @@
 .Box-header {
   display: flex;
   flex-direction: row;
-  justify-content: space-between;
+  justify-content: flex-start;
   align-items: center;
   padding: 0 25px;
   margin: -1px -1px 0;
@@ -34,13 +34,15 @@
   border-top-right-radius: 6px;
 }
 
-.Box-header svg {
+.Box-header svg.tooltip-icon {
   cursor: help;
+  margin-top: 4px;
 }
 
 .Box-title {
   font-size: 18px;
   font-weight: 600;
+  margin-right: 8px;
 }
 
 ul {
